refactor: take core::time::Duration for toast timeout
The toast-timeout parameter was typed as embassy_time::Duration but the
value is immediately collapsed to a u32 millisecond count internally, so
the type was pure ceremony that forced every consumer of a timed toast to
add an embassy-time dependency just to express a timeout.

Switch the public toast API (Navigator::show_toast, navigator::post_toast,
NavAction::ShowToast / show_toast) to the dependency-free, no_std
core::time::Duration. embassy-time stays an internal dependency for the
render-loop timers and tick source, but no longer appears in the public
toast signature, so it no longer propagates to consumers.

No behavior change; the millisecond saturation cast is unchanged apart from
core::time::Duration::as_millis() returning u128. 0.3.1 was never published,
so this is folded into a 0.3.2 release rather than a breaking bump.

+++ src/navigator.rs
@@ -17,7 +17,7 @@ use alloc::collections::VecDeque;
 use alloc::vec::Vec;
 
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, channel::Channel};
-use embassy_time::Duration;
+use core::time::Duration;
 use oxivgl_sys::*;
 
 use crate::driver::get_tick_ms;
@@ -588,11 +588,11 @@ impl Navigator {
         self.toast = Some(boxed);
         self.toast_container = Some(container);
         self.toast_deadline_ms = duration.map(|d| {
-            // Saturate the embassy Duration into u32 ms (≈49.7 days max),
-            // then wrap-add to the current tick. The compare in tick_toast
-            // uses `wrapping_sub` so wrap-around is correct as long as the
+            // Saturate the Duration into u32 ms (≈49.7 days max), then
+            // wrap-add to the current tick. The compare in tick_toast uses
+            // `wrapping_sub` so wrap-around is correct as long as the
             // duration is < ~25 days.
-            let ms = d.as_millis().min(u32::MAX as u64) as u32;
+            let ms = d.as_millis().min(u32::MAX as u128) as u32;
             get_tick_ms().wrapping_add(ms)
         });
         true
